expose wrapped Acorn methods which include comment information (#26)
